add context to io errors
