+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/qmuntal/opc
 
 require github.com/stretchr/testify v1.3.0
+
+go 1.13
