+++ relationship.go
@@ -2,6 +2,7 @@ package opc
 
 import (
 	"encoding/xml"
+	"fmt"
 	"io"
 	"math/rand"
 	"net/url"
@@ -156,10 +157,10 @@ func encodeRelationships(w io.Writer, rs []*Relationship) error {
 	return xml.NewEncoder(w).Encode(re)
 }
 
-func decodeRelationships(r io.Reader) ([]*Relationship, error) {
+func decodeRelationships(r io.Reader, partName string) ([]*Relationship, error) {
 	relDecode := new(relationshipsXML)
 	if err := xml.NewDecoder(r).Decode(relDecode); err != nil {
-		return nil, err
+		return nil, fmt.Errorf("opc: %s: cannot be decoded: %w", partName, err)
 	}
 	rel := make([]*Relationship, len(relDecode.RelsXML))
 	for i, rl := range relDecode.RelsXML {
